+++ tests/corelib_2/int_parse_with_limited_ints_test.dart
@@ -9,9 +9,6 @@
 import "package:expect/expect.dart";
 
 main() {
-  const int ERROR = 42;
-  var returnError = (_) => ERROR;
-
   Expect.equals(0, int.parse("0"));
   Expect.equals(1, int.parse("1"));
   Expect.equals(-1, int.parse("-1"));
@@ -28,11 +25,11 @@ main() {
   Expect.equals(-9223372036854775807, int.parse("-9223372036854775807"));
   Expect.equals(-9223372036854775807 - 1, int.parse("-9223372036854775808"));
 
-  Expect.equals(ERROR, int.parse("-0x8000000000000001", onError: returnError));
-  Expect.equals(ERROR, int.parse("9223372036854775808", onError: returnError));
-  Expect.equals(ERROR, int.parse("9223372036854775809", onError: returnError));
-  Expect.equals(ERROR, int.parse("-9223372036854775809", onError: returnError));
-  Expect.equals(ERROR, int.parse("10000000000000000000", onError: returnError));
+  Expect.throws(() => int.parse("-0x8000000000000001"));
+  Expect.throws(() => int.parse("9223372036854775808"));
+  Expect.throws(() => int.parse("9223372036854775809"));
+  Expect.throws(() => int.parse("-9223372036854775809"));
+  Expect.throws(() => int.parse("10000000000000000000"));
 
   Expect.equals(
       0x7fffffffffffffff,
@@ -50,28 +47,16 @@ main() {
           "-1000000000000000000000000000000000000000000000000000000000000000",
           radix: 2));
 
-  Expect.equals(
-      ERROR,
-      int.parse(
-          "1000000000000000000000000000000000000000000000000000000000000000",
-          radix: 2,
-          onError: returnError));
-  Expect.equals(
-      ERROR,
-      int.parse(
-          "1111111111111111111111111111111111111111111111111111111111111110",
-          radix: 2,
-          onError: returnError));
-  Expect.equals(
-      ERROR,
-      int.parse(
-          "1111111111111111111111111111111111111111111111111111111111111111",
-          radix: 2,
-          onError: returnError));
-  Expect.equals(
-      ERROR,
-      int.parse(
-          "-1000000000000000000000000000000000000000000000000000000000000001",
-          radix: 2,
-          onError: returnError));
+  Expect.throws(() => int.parse(
+      "1000000000000000000000000000000000000000000000000000000000000000",
+      radix: 2));
+  Expect.throws(() => int.parse(
+      "1111111111111111111111111111111111111111111111111111111111111110",
+      radix: 2));
+  Expect.throws(() => int.parse(
+      "1111111111111111111111111111111111111111111111111111111111111111",
+      radix: 2));
+  Expect.throws(() => int.parse(
+      "-1000000000000000000000000000000000000000000000000000000000000001",
+      radix: 2));
 }